+++ common/components/News/NewsContent/index.js
@@ -1,6 +1,5 @@
 export FontSize from './FontSize';
 export Content from './Content';
-export HotVideoBlocks from './HotVideoBlocks';
 export RecommendAds from './RecommendAds';
 export RelatedContent from './RelatedContent';
 export Social from './Social';