+++ common/routes/Video/components/BlockItems9.js
@@ -20,11 +20,10 @@ class BlockItems9 extends PureComponent {
   render () {
     const { local, newsList, page } = this.props;
     const items = newsList.map(({ sn, MainPhoto, shortTitle, startedAt }, index) => {
-      const photo = MainPhoto || {};
       return (
         <div key={sn} className={css(styles.blockItem)} onClick={() => { this.select(index); }}>
           <TypeIcon type='VIDEO' />
-          <img className={css(styles.img)} src={photo.googleCDN || photo.thumbnail || photo.url} />
+          <img className={css(styles.img)} src={MainPhoto.thumbnail} />
           <div className={css(styles.bottom)}>
             <div className={css(styles.title)}><h3 className={css(styles.h3)}>{shortTitle}</h3></div>
             <img src='/icons/whiteClock.png' />