+++ server/sitemap.js
@@ -84,6 +84,5 @@ module.exports = function (app) {
   return function (req, res, next) {
     return next();
   };
-
 };